requirements.txt
numpy==1.13.3+mkl
colorthief==0.2.1
praw==5.4.0
Pillow==5.3.0
